backend/requirements.txt
    # Core
    fastapi
    uvicorn
    python-multipart
    pydantic
    starlette

    # Firebase
    firebase-admin
    google-cloud-firestore
    google-cloud-storage

    # Google Vision
    google-cloud-vision

    # Utilities
    python-dotenv
    python-jose
    uuid

    # Production Server
    uvicorn
    openai
    tenacity
    pdf2image
    pillow
    google-generativeai
    google-auth
    numpy